requirements.txt
selenium>=4.22.0
webdriver-manager>=4.0.2
pillow>=10.3.0